script: Move globalscope interfaces into globalscope/
Part of #38901

Testing: it compiles
