+++ components/script/dom/globalscope/origin.rs
