+++ components/script/dom/globalscope/script_execution.rs
