+++ components/script/dom/mod.rs
@@ -274,8 +274,8 @@ pub(crate) mod geolocation;
 pub(crate) use self::geolocation::*;
 pub(crate) mod geometry;
 pub(crate) use self::geometry::*;
-pub(crate) mod global_scope_script_execution;
 pub(crate) mod globalscope;
+pub(crate) use self::globalscope::*;
 pub(crate) mod html;
 pub(crate) use self::html::*;
 pub(crate) mod indexeddb;
@@ -284,9 +284,6 @@ pub(crate) mod intersectionobserver;
 pub(crate) mod intersectionobserverentry;
 pub(crate) mod media;
 pub(crate) use self::media::*;
-pub(crate) mod messagechannel;
-#[expect(dead_code)]
-pub(crate) mod messageport;
 pub(crate) mod mimetype;
 pub(crate) mod mimetypearray;
 pub(crate) mod mutationobserver;
@@ -297,7 +294,6 @@ pub(crate) use self::navigator::*;
 pub(crate) mod node;
 pub(crate) use self::node::*;
 pub(crate) mod notification;
-pub(crate) mod origin;
 pub(crate) mod performance;
 pub(crate) use self::performance::*;
 pub(crate) mod permissions;
